tree: rename nvm_child_buffer
